+++ extensions/chatgpt/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Fix] - 2024-1-19
+
+- Fix: Azure API 401 error
+
 ## [Chore] - 2024-01-11
 
 - Migrate openai v3 to v4